add generic patch macrotask method
